Success on making workspace, floating action button, floating workspace menu separate, reusable components that I can pass props to
