+++ src/components/FloatingControls.js
@@ -1,0 +1,29 @@
+import React, { Component } from 'react';
+import {
+  Button,
+  Dropdown,
+  Icon
+} from 'semantic-ui-react';
+
+import './components.css';
+
+class FloatingActionButton extends Component {
+  render () {
+    return (
+      <Button className='FloatingActionButton' circular color={this.props.color} size='huge' icon={this.props.icon} />
+    )
+  }
+}
+
+class FloatingWorkspaceMenu extends Component {
+  render () {
+    return (
+      <Dropdown icon={<Icon name='ellipsis vertical' size='big' color='grey' />} style={{ position: 'fixed', top: '2rem', right: '1rem' }} >
+        <Dropdown.Menu direction='left' children={this.props.children}>
+        </Dropdown.Menu>
+      </Dropdown>
+    )
+  }
+}
+
+export { FloatingActionButton, FloatingWorkspaceMenu }
