+++ src/components/components.css
@@ -4,3 +4,13 @@
 	right: 2rem;
 	box-shadow: 0 6px 10px 0 rgba(0,0,0,0.14), 0 1px 18px 0 rgba(0,0,0,0.12), 0 3px 5px -1px rgba(0,0,0,0.2)!important;
 }
+
+.WorkspaceGrid {}
+
+.WorkspaceGridRow {
+	margin-left: 14rem;
+}
+
+.WideColumnWorkspace {}
+
+.NarrowColumnWorkspace {}
