+++ src/containers/Books/Controls.js
@@ -9,7 +9,7 @@ import {
   Menu,
   Select
 } from 'semantic-ui-react';
-import FloatingActionButton from '../../components/FloatingActionButton'
+import { FloatingActionButton } from '../../components/FloatingControls'
 
 import './Books.css';
 
@@ -18,7 +18,7 @@ class BooksControls extends Component {
     return (
       <div>
         {/* <Button circular color='blue' size='huge' icon='play' style={{ position: 'fixed', bottom: '7rem', right: '2rem', boxShadow: '0 6px 10px 0 rgba(0,0,0,0.14), 0 1px 18px 0 rgba(0,0,0,0.12), 0 3px 5px -1px rgba(0,0,0,0.2)' }} /> */}
-        <FloatingActionButton />
+        <FloatingActionButton icon='plus' color='google plus' />
       </div>
     );
   };
