+++ src/containers/Tables/Controls.js
@@ -8,7 +8,7 @@ import {
   Label,
   Menu
 } from 'semantic-ui-react';
-import FloatingActionButton from '../../components/FloatingActionButton'
+import { FloatingActionButton, FloatingWorkspaceMenu } from '../../components/FloatingControls'
 
 import './Tables.css';
 
@@ -16,14 +16,12 @@ class TablesControls extends Component {
   render() {
     return (
       <div>
-        <Dropdown icon={<Icon name='ellipsis vertical' size='big' color='grey' />} style={{ position: 'fixed', top: '2rem', right: '1rem' }} >
-            <Dropdown.Menu direction='left'>
-              <Dropdown.Item text='Import table[s] ...' />
-              <Dropdown.Item text='Export tables ...' />
-              <Dropdown.Item text='Delete all tables' />
-            </Dropdown.Menu>
-          </Dropdown>
-        <FloatingActionButton />
+        <FloatingWorkspaceMenu>
+          <Dropdown.Item text='Import table[s] ...' />
+          <Dropdown.Item text='Export tables ...' />
+          <Dropdown.Item text='Delete all tables' />
+        </FloatingWorkspaceMenu>
+        <FloatingActionButton icon='plus' color='google plus' />
       </div>
     );
   };
